Feature complete

+++ vs_server.h
@@ -24,17 +24,17 @@ public:
     u_int32_t getIp();
     SERVER_STATE getState();
     u_int8_t getTTD();
-    u_int16_t getBlockId();
+    u_int64_t getBlockId();
     void setState(SERVER_STATE state);
     void refreshTTD();
     void decrementTTD();
-    void setBlockId(u_int16_t block_id);
+    void setBlockId(u_int64_t block_id);
 
 private:
     u_int32_t server_ip;
     SERVER_STATE server_status;
     u_int8_t time_to_death;
-    u_int16_t block_id;
+    u_int64_t block_id;
 };
 
 #endif // VS_SERVER_H
